Refactor itinerary map component by extracting utility functions for route saving and toggle expansion
